requirements.txt
aioconsole==0.1.15
aiogram==2.3
aiohttp==3.6.1
aiohttp-socks==0.2.2
aiohttp-jinja2==1.2.0
aiohttp-swagger==1.0.9
pyyaml==5.1
aio-pika==5.6.0
motor==2.0.0
tqdm==4.36.1
pymongo==3.10.1
sentry-sdk==0.19.5
pydantic==1.8.2
hydra-core==1.1.1
